+++ src/authentication/index.js
@@ -47,6 +47,127 @@ module.exports = function(router) {
   };
 
   /**
+   * Checks to see if a decoded token is allowed to access this path.
+   * @param req
+   * @param decoded
+   * @return {boolean}
+   */
+  var isTokenAllowed = function(req, decoded) {
+    if (!decoded.allow) {
+      return true;
+    }
+
+    var isAllowed = false;
+    var allowed = decoded.allow.split(',');
+    _.each(allowed, function(allow) {
+      var parts = allow.split(':');
+      if (parts.length < 2) {
+        return;
+      }
+
+      // Make sure the methods match.
+      if (req.method.toUpperCase() !== parts[0].toUpperCase()) {
+        return;
+      }
+
+      try {
+        var regex = new RegExp(parts[1]);
+        if (regex.test(req.baseUrl + req.url)) {
+          isAllowed = true;
+          return false;
+        }
+      }
+      catch (err) {
+        console.warn('Bad token allow string.');
+      }
+    });
+    return isAllowed;
+  };
+
+  /**
+   * Generate a temporary token for a specific path and expiration.
+   *
+   * @param token
+   * @param allow
+   * @param expire
+   * @param cb
+   */
+  var getTempToken = function(req, res, token, allow, expire, cb) {
+    // Decode/refresh the token and store for later middleware.
+    jwt.verify(token, router.formio.config.jwt.secret, function(err, decoded) {
+      if (err || !decoded) {
+        if (err && (err.name === 'JsonWebTokenError')) {
+          return cb('Bad Token');
+        }
+        else if (err && (err.name === 'TokenExpiredError')) {
+          return cb('Login Timeout');
+        }
+        else {
+          return cb('Unauthorized');
+        }
+      }
+
+      if (!res.headersSent) {
+        res.setHeader('Access-Control-Expose-Headers', 'x-jwt-token');
+        res.setHeader('x-jwt-token', token);
+      }
+
+      // Check to see if this path is allowed.
+      if (!isTokenAllowed(req, decoded)) {
+        return res.status(401).send('Token path not allowed.');
+      }
+
+      // Do not allow regeneration of temp tokens from other temp tokens.
+      if (decoded.tempToken) {
+        return cb('Cannot issue a temporary token using another temporary token.');
+      }
+
+      var now = Math.round((new Date()).getTime() / 1000);
+      expire = expire || 3600;
+      expire = parseInt(expire, 10);
+      var timeLeft = (parseInt(decoded.exp, 10) - now);
+
+      // Ensure they are not trying to create an extended expiration.
+      if ((expire > 3600) && (timeLeft < expire)) {
+        return cb('Cannot generate extended expiring temp token.')
+      }
+
+      // Add the allowed to the token.
+      if (allow) {
+        decoded.allow = allow;
+      }
+
+      decoded.tempToken = true;
+      jwt.sign(decoded, router.formio.config.jwt.secret, {
+        expiresIn: expire
+      }, (token) => cb(null, token));
+    });
+  };
+
+  var tempToken = function(req, res, next) {
+    if (!req.headers) {
+      return res.status(400).send('No headers provided.');
+    }
+
+    var token = req.headers['x-jwt-token'];
+    var allow = req.headers['x-allow'];
+    var expire = req.headers['x-expire'];
+    if (!token) {
+      return res.status(400).send('You must provide an existing token in the x-jwt-token header.');
+    }
+
+    // get a temporary token.
+    getTempToken(req, res, token, allow, expire, function(err, tempToken) {
+      if (err) {
+        return res.status(400).send(err);
+      }
+
+      // Send the temp token as a response.
+      return res.send(tempToken);
+    });
+  };
+
+  /**
    * Authenticate a user.
    *
    * @param forms {Mixed}
@@ -207,8 +328,11 @@ module.exports = function(router) {
    */
   return {
     getToken: getToken,
+    isTokenAllowed: isTokenAllowed,
+    getTempToken: getTempToken,
     authenticate: authenticate,
     currentUser: currentUser,
+    tempToken: tempToken,
     logout: function(req, res) {
       res.setHeader('x-jwt-token', '');
       res.sendStatus(200);