+++ Changelog.md
@@ -13,6 +13,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 
+## 1.16.4
+### Added
+ - Added temporary auth token support. You can now generate new tokens using the
+   /token endpoint with the following headers.
+    - x-expire = The expiration of the token in seconds.
+    - x-allow = The paths to allow for the token in the format: GET:/path/[0-9a-z]+
+
 ## 1.16.3
 ### Added
  - Make action handlers alterable.